refactor: remove re-export of db config via database service
