+++ containers/ecr-viewer/integration/listEcrDataService.test.ts
@@ -8,9 +8,10 @@ import {
   SqlBool,
 } from "kysely";
 
-import { dbDialect, dbNamespace, getDb } from "@/app/api/services/database";
+import { getDb } from "@/app/api/services/database";
 import { Core, NewCoreECR } from "@/app/api/services/types/core";
 import { NewExtendedECR } from "@/app/api/services/types/extended";
+import { dbDialect, dbNamespace } from "@/app/data/db/utils/db-config";
 import { formatDate, formatDateTime } from "@/app/services/formatDateService";
 import {
   CoreMetadataModel,
